[Commons] Added SelcukesFixer
