+++ selcukes-commons/src/main/java/io/github/selcukes/commons/listener/LifecycleManager.java
@@ -16,7 +16,7 @@
 
 package io.github.selcukes.commons.listener;
 
-import io.github.selcukes.commons.result.TestResult;
+import io.github.selcukes.commons.fixer.TestResult;
 import lombok.CustomLog;
 
 import java.util.List;
