+++ selcukes-core/src/main/java/io/github/selcukes/core/validation/Validation.java
@@ -22,7 +22,7 @@ import lombok.experimental.UtilityClass;
 import java.util.ArrayList;
 import java.util.List;
 
-import static io.github.selcukes.commons.result.Validator.fail;
+import static io.github.selcukes.commons.fixer.SelcukesFixer.fail;
 
 @UtilityClass
 @CustomLog
